relocates pawn promotion logic to moves file
