+++ src/utils/moves/moves.js
@@ -1,6 +1,7 @@
 import { produce } from 'immer';
 import { getPieceAtSquare, validateSquare } from '../board';
 import { DevError } from '../errors';
+import config from '../../config/config';
 
 import kingMove from './king';
 import queenMove from './queen';
@@ -42,16 +43,34 @@ function excludeOccupiedSquares(squares, board, color) {
   });
 }
 
+const backRank = {
+  w: config.get('board.dimensions.numberRanks'),
+  b: 1,
+};
+const promotionPieces = ['q', 'r', 'b', 'n'];
 export function makeMove(board, start, end) {
   validateSquare(start);
   validateSquare(end);
 
-  const piece = getPieceAtSquare(board, start);
+  let piece = getPieceAtSquare(board, start);
   if (!piece)
     throw new DevError(`No piece at start square ${JSON.stringify(start)}`);
+
+  if (piece.type === 'p' && end.rank === backRank[piece.color]) {
+    piece = promotePawn(piece.color);
+  }
 
   return produce(board, (draft) => {
     draft[end.rank][end.file] = piece;
     draft[start.rank][start.file] = undefined;
   });
+}
+
+function promotePawn(color) {
+  let promotionPiece;
+  do {
+    promotionPiece = prompt(`Promote to: (${promotionPieces.join(', ')})`);
+  } while (!promotionPieces.includes(promotionPiece));
+
+  return { type: promotionPiece, color };
 }
